+++ packages/table/src/ExpandedContent/ExpandedContent.tsx
@@ -15,7 +15,9 @@ const ExpandedContent = <T extends RowData>({
   row,
 }: ExpandedContentProps<T>) => {
   const isExpanded = row.getIsExpanded();
-  const Content = row?.original?.renderExpandedContent && row?.original?.renderExpandedContent(row);
+  const Content =
+    row?.original?.renderExpandedContent &&
+    row?.original?.renderExpandedContent(row);
 
   const { theme } = useDarkMode();
   return (