+++ packages/table/src/ExpandedContent/ExpandedContent.types.ts
@@ -1,6 +1,6 @@
 import { RowData } from '@tanstack/react-table';
 
-import { LeafyGreenTableRow } from '../useLeafygreenTable';
+import { LeafyGreenTableRow } from '../useLeafyGreenTable';
 
 // type WithRequired<T, K extends keyof T> = T & { [P in K]-?: T[P] };
 export interface ExpandedContentProps<T extends RowData> {