+++ packages/table/src/Row/Row.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { LGRowData } from '../useLeafygreenTable';
+import { LGRowData } from '../useLeafyGreenTable';
 
 import InternalRowWithoutRT from './InternalRowWithoutRT';
 import InternalRowWithRT from './InternalRowWithRT';
@@ -10,7 +10,11 @@ import { RowProps } from './Row.types';
 /**
  * Renders the provided cells
  */
-const Row = <T extends LGRowData>({ row, virtualRow, ...rest }: RowProps<T>) => {
+const Row = <T extends LGRowData>({
+  row,
+  virtualRow,
+  ...rest
+}: RowProps<T>) => {
   if (row) {
     return <InternalRowWithRT row={row} virtualRow={virtualRow} {...rest} />;
   } else {