+++ packages/table/src/Row/Row.types.ts
@@ -3,7 +3,7 @@ import { VirtualItem } from 'react-virtual';
 
 import { HTMLElementProps } from '@leafygreen-ui/lib';
 
-import { LeafyGreenTableRow, LGRowData } from '../useLeafygreenTable';
+import { LeafyGreenTableRow, LGRowData } from '../useLeafyGreenTable';
 
 export interface InternalRowBaseProps extends HTMLElementProps<'tr'> {
   /**
@@ -17,11 +17,11 @@ export interface InternalRowWithoutRTProps extends InternalRowBaseProps {}
 export interface InternalRowWithRTProps<T extends LGRowData>
   extends InternalRowBaseProps {
   /**
-   * Row object passed from the `useLeafygreenTable` hook.
+   * Row object passed from the `useLeafyGreenTable` hook.
    */
   row: LeafyGreenTableRow<T>;
   /**
-   * Virtual row object passed from the `useLeafygreenTable` hook
+   * Virtual row object passed from the `useLeafyGreenTable` hook
    */
   virtualRow?: VirtualItem;
   isNestedRow?: boolean;