+++ packages/table/src/Table.story.tsx
@@ -19,11 +19,13 @@ import Row from './Row/Row';
 import TableBody from './TableBody/TableBody';
 import TableContainer from './TableContainer/TableContainer';
 import TableHead from './TableHead/TableHead';
-import useLeafygreenTable from './useLeafygreenTable/useLeafygreenTable';
 import { makeData, Person } from './utils/makeData';
 import { AnyDict } from './utils/types';
 import Table from './Table';
-import { LeafyGreenTableCell, LeafyGreenTableRow } from './useLeafygreenTable';
+import useLeafyGreenTable, {
+  LeafyGreenTableCell,
+  LeafyGreenTableRow,
+} from './useLeafyGreenTable';
 
 export default {
   title: 'Components/Table',
@@ -124,7 +126,7 @@ export const NestedRows: ComponentStory<typeof Table> = () => {
     [],
   );
 
-  const table = useLeafygreenTable<Person>({
+  const table = useLeafyGreenTable<Person>({
     containerRef: tableContainerRef,
     data,
     columns,
@@ -229,7 +231,7 @@ export const ExpandableContent: ComponentStory<typeof Table> = args => {
     [],
   );
 
-  const table = useLeafygreenTable<Person>({
+  const table = useLeafyGreenTable<Person>({
     containerRef: tableContainerRef,
     data,
     columns,
@@ -334,7 +336,7 @@ export const SortableRows: ComponentStory<typeof Table> = args => {
     [],
   );
 
-  const table = useLeafygreenTable<Person>({
+  const table = useLeafyGreenTable<Person>({
     containerRef: tableContainerRef,
     data,
     columns,
@@ -441,7 +443,7 @@ export const SelectableRows: ComponentStory<typeof Table> = args => {
     [],
   );
 
-  const table = useLeafygreenTable<Person>({
+  const table = useLeafyGreenTable<Person>({
     containerRef: tableContainerRef,
     data,
     columns,
@@ -566,7 +568,7 @@ export const WithPagination: ComponentStory<typeof Table> = args => {
     [],
   );
 
-  const table = useLeafygreenTable<Person>({
+  const table = useLeafyGreenTable<Person>({
     containerRef: tableContainerRef,
     data,
     columns,