+++ packages/table/src/TableBody/TableBody.types.ts
@@ -1,6 +1,6 @@
 import { HTMLElementProps } from '@leafygreen-ui/lib';
 
-import { LGRowData } from '../useLeafygreenTable';
+import { LGRowData } from '../useLeafyGreenTable';
 
 export type TableBodyProps<T extends LGRowData> = {
   /**